extract getRandomMatriculationId
